add logic for expired discounts during purchases and add Active discounts model to track the user discounts

+++ src/pages/api/order/add.ts
@@ -1,7 +1,8 @@
 import connectToDB from "@/config/db";
-import { unknownObjProps, userRelatedDataTypes } from "@/global.t";
+import ActiveDiscountModel from "@/models/Discount/ActiveDiscount";
+import DiscountModel from "@/models/Discount/Discount";
 import { transactionModel } from "@/models/Transactions";
-import { BasketItemModel, NotificationModel, OrderModel } from "@/models/UserRelatedSchemas";
+import { BasketItemModel, NotificationModel } from "@/models/UserRelatedSchemas";
 import { NextApiRequest, NextApiResponse } from "next";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
@@ -18,9 +19,21 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 
         if (!userOrders.length) return res.status(422).json({ message: 'محصولی برای سفارش وجود نداره😂' })
 
-        const userOrdersPlain = userOrders.map((order: any) => order.toObject()); // use toObject so we can see the populated products data in client
+        const userOrdersPlain = userOrders.map((order: any) => order.toObject()); // use toObject so we can see the populated products data in client (we can use .lean() too)
+
+        const checkForDiscount = await ActiveDiscountModel.findOne({ userID, isUsed: false })
+
+        if (checkForDiscount) {
+
+            const discountTarget = await DiscountModel.findOne({ code: checkForDiscount.code })
+            if (discountTarget.maxUse <= 0) return res.status(421).json({ message: 'کد تخفیف شما منقضی شده. برای ادامه خرید کد تخفیف را حذف کنید' })
+
+            await DiscountModel.findOneAndUpdate({ code: checkForDiscount.code }, { maxUse: discountTarget.maxUse - 1 })
+            await ActiveDiscountModel.findOneAndUpdate({ userID, isUsed: false }, { isUsed: true }) // so user can't use this discount anymore
+        }
 
         const newOrderTransaction = await transactionModel.create({ productsList: userOrdersPlain, userID, customerData, totalPrice, status: 'PROCESSING' })
+
 
         await BasketItemModel.deleteMany({ userID }) // clear the user basket
